// frontend/src/api/backend/backendSchemas.ts
export type ObjectId = {};
// ...
export type ModulePermissionResponseDto = {
  /**
   * Id of the permission
   *
   * @example  677d567870dc1af598b322a7
   */
  _id: ObjectId;
  /**
   * Name of the permission
   *
   * @example Contract
   */
  name: string;
};

export type ModuleResponseDto = {
  /**
   * Id of the module
   *
   * @example  677d567870dc1af598b322a7
   */
  _id: ObjectId;
  /**
   * Name of the module
   *
   * @example Contract
   */
  name: string;
  /**
   * Name of the module
   */
  permissions: ModulePermissionResponseDto[];
};